+++ pyrefly/lib/lsp/non_wasm/mod.rs
@@ -7,6 +7,7 @@
 
 mod build_system;
 pub mod call_hierarchy;
+pub mod convert_module_package;
 pub mod document_symbols;
 pub mod folding_ranges;
 pub mod lsp;
